create directory for scripts
